+++ CHANGELOG.md
@@ -10,6 +10,9 @@ toc: hidden
 
 - Feat
   - 1、新增`drop-select-item`组件点击可阻断方法
+  - 2、新增`icon`组件`icon`、`iconImg`属性，可以传入gif等的图片
+  - 3、新增`toast`组件`loading`方法`iconImg`属性，可以传入gif等的图片
+  - 4、调整`toast`组件加载loading位置由左右布局改为上下布局
 
 - Fix
   - 1、修复`drop-select-item`组件`pickerTitle`属性失效问题。